+++ app/models/concerns/lcms/engine/navigable.rb
@@ -13,29 +13,25 @@ module Lcms
         end
 
         def previous
-          @previous ||= begin
-            return unless level_position
-
-            if level_position.positive?
+          @previous ||=
+            if level_position.to_i.positive?
               siblings.where(level_position: level_position - 1).first
             else
               # last element of previous node from parent level
               parent.try(:previous).try(:children).try(:last)
             end
-          end
         end
 
         def next
-          @next ||= begin
-            return unless level_position
-
-            if level_position < siblings.size
+          @next ||=
+            if level_position.nil?
+              nil
+            elsif level_position < siblings.size
               siblings.where(level_position: level_position + 1).first
             else
               # first element of next node from parent level
               parent.try(:next).try(:children).try(:first)
             end
-          end
         end
       end
     end