+++ spec/factories/documents.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
 FactoryBot.define do
-  factory :document, class: Lcms::Engine::Document do
-  end
+  factory :document, class: Lcms::Engine::Document
 end