+++ spec/factories/downloads.rb
@@ -2,7 +2,7 @@
 
 FactoryBot.define do
   factory :download, class: Lcms::Engine::Download do
-    title { Faker::Lorem.words.join '' }
+    title { Faker::Lorem.words.join }
     url { Faker::Internet.url }
   end
 end